Split debug components into separate files
Extract DebugGameView and GameScreenForPlayer into their own component
files; page.tsx retains DebugPage and imports from them.
